fix(sana-wm): keep GDN chunkwise B-kernel fp32 under torch.compile
`_phase_a_z_kernel` computed `K = tl.where(...) * k_scale`, where k_scale
(K_SCALE) is a Python-float kernel arg. Inductor's user-defined-Triton-kernel
path types that scalar as fp64 (eager keeps it fp32), so `beta_K -> tl.sum`
promoted the fp32 loop-carried accumulator `B_acc` to fp64 and Triton rejected
the loop-carried type change — torch.compile of the SanaWM DiT failed (and under
the default max-autotune mode stalled in the inductor subprocess compile pool
until the scheduler's 100-min timeout).

Cast K back to fp32 after the scale (no-op in eager; matches eager numerics).
With this, torch.compile works and DiT denoise drops from 3.17 s/it (eager) to
~1.98 s/it steady on H200.

+++ python/sglang/jit_kernel/diffusion/triton/sana_wm_gdn_chunkwise.py
@@ -485,7 +485,11 @@ def _phase_a_z_kernel(
             k_inv_rms_ptr + pid_b * N + n_idx, mask=mask_s, other=1.0
         ).to(tl.float32)
         K_normed = K_raw * k_inv_rms[:, None] * k_nw[None, :]
-        K = tl.where(K_normed > 0, K_normed, 0.0) * k_scale
+        # Cast back to fp32: K_SCALE is a Python-float kernel arg, which inductor's
+        # user-defined-Triton-kernel path types as fp64 (eager keeps it fp32). Without
+        # this, beta_K -> tl.sum promotes the fp32 loop-carried B_acc to fp64 and Triton
+        # rejects the type change under torch.compile. No-op in eager.
+        K = (tl.where(K_normed > 0, K_normed, 0.0) * k_scale).to(tl.float32)
 
         beta_K = beta_t[:, None] * K
